refactor(backend): unify sync/async MySQL backend parity and fix execute double definition
- Merge two `execute` definitions into one with both options construction
  and retry logic, fixing the bug where retry logic was dead code due to
  Python class semantics overriding the first definition with the second
- Move non-I/O methods (CONNECTION_ERROR_CODES, _is_connection_error,
  _handle_error) to shared MySQLBackendMixin
- Rewrite transaction classes to inherit MySQLTransactionMixin, removing
  duplicated _build_set_isolation_sql
- Change async _handle_error from async def to def (pure error
  classification, no I/O)
- Fix minor sync/async inconsistencies in __init__, ping,
  get_server_version
- Fix test import in test_async_error_handling.py
